+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/FlutterWebView.java
@@ -176,9 +176,10 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
         .setJavaScriptCanOpenWindowsAutomatically(
             true) // Always allow automatically opening of windows.
         .setSupportMultipleWindows(true) // Always support multiple windows.
-        .setWebChromeClient(webChromeClient)
-        .setDownloadListener(
-            downloadListener); // Always use {@link FlutterWebChromeClient} as web Chrome client.
+        .setWebChromeClient(
+            webChromeClient) // Always use {@link FlutterWebChromeClient} as web Chrome client.
+        .setDownloadListener(downloadListener)
+        .setZoomControlsEnabled(true); // Always use built-in zoom mechanisms.
 
     return webViewBuilder.build();
   }
@@ -244,7 +245,11 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
         currentUrl(result);
         break;
       case "evaluateJavascript":
-        evaluateJavaScript(methodCall, result);
+      case "runJavascriptReturningResult":
+        evaluateJavaScript(methodCall, result, true);
+        break;
+      case "runJavascript":
+        evaluateJavaScript(methodCall, result, false);
         break;
       case "addJavascriptChannels":
         addJavaScriptChannels(methodCall, result);
@@ -325,7 +330,8 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
   }
 
   @TargetApi(Build.VERSION_CODES.KITKAT)
-  private void evaluateJavaScript(MethodCall methodCall, final Result result) {
+  private void evaluateJavaScript(
+      MethodCall methodCall, final Result result, final boolean returnValue) {
     String jsString = (String) methodCall.arguments;
     if (jsString == null) {
       throw new UnsupportedOperationException("JavaScript string cannot be null");
@@ -335,7 +341,11 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
         new android.webkit.ValueCallback<String>() {
           @Override
           public void onReceiveValue(String value) {
-            result.success(value);
+            if (returnValue) {
+              result.success(value);
+            } else {
+              result.success(null);
+            }
           }
         });
   }
@@ -428,6 +438,9 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
         case "allowsInlineMediaPlayback":
           // no-op inline media playback is always allowed on Android.
           break;
+        case "zoomEnabled":
+          setZoomEnabled((boolean) settings.get(key));
+          break;
         default:
           throw new IllegalArgumentException("Unknown WebView setting: " + key);
       }
@@ -465,6 +478,10 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
 
   private void updateUserAgent(String userAgent) {
     webView.getSettings().setUserAgentString(userAgent);
+  }
+
+  private void setZoomEnabled(boolean shouldEnable) {
+    webView.getSettings().setSupportZoom(shouldEnable);
   }
 
   @Override