+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.2.0
+
+* Implemented new `runJavascript` and `runJavascriptReturningResult` methods in platform interface.
+
+## 2.1.0
+
+* Add `zoomEnabled` functionality.
+
 ## 2.0.15
 
 * Added Overrides in  FlutterWebView.java